+++ src/environments/environment.test.ts
@@ -12,8 +12,8 @@ export const environment = {
   useRegionPerLocale: false,
   useFeedbackPrompt: true,
 
+  useSearchApi: false,
   searchApi: '',
-  searchApiKey: '',
 
   // Configuration of content
   localeLanguage: 'en',
@@ -47,5 +47,6 @@ export const environment = {
 
   // Monitoring/Telemetry:
   matomoConnectionString: '',
+  goatCounterConnectionString: '',
   appInsightsConnectionString: '',
 };